feat: adicionando helper do parse de dados e interface do status de transacoes
