+++ src/modules/transaction/interfaces/ITransaction.ts
@@ -2,7 +2,7 @@ export interface IState {
 	id: string
 	title: string
 	description: string
-	status: string
+	status: 'cancelled' | 'created' | 'processing' | 'processed';
 	amount: number
 	date: Date
 	from: string
